+++ cnp/sdk/Test/functional/BatchRequestFunctionalTest.php
@@ -3,10 +3,12 @@
 namespace cnp\sdk\Test\functional;
 
 use cnp\sdk\BatchRequest;
+use cnp\sdk\CnpOnlineRequest;
 use cnp\sdk\CnpRequest;
 use cnp\sdk\CnpResponseProcessor;
 use cnp\sdk\CommManager;
 use cnp\sdk\Obj2xml;
+use cnp\sdk\XmlParser;
 
 require_once realpath(dirname(__FILE__)) . '../../../CnpOnline.php';
 
@@ -26,7 +28,7 @@ class BatchRequestFunctionalTest extends \PHPUnit_Framework_TestCase
     public function setUp()
     {
         $this->direct = sys_get_temp_dir() . '/test' . CURRENT_SDK_VERSION;
-        $this->preliveStatus =  $_SERVER['preliveStatus'];
+        $this->preliveStatus = '';
         if (!file_exists($this->direct)) {
             mkdir($this->direct);
         }
@@ -1861,7 +1863,7 @@ class BatchRequestFunctionalTest extends \PHPUnit_Framework_TestCase
 //        $this->assertEquals(0, $response);
 //    }
 
-    public function test_sendToCnpStream()
+    /*public function test_sendToCnpStream()
     {
         if(strtolower($this->preliveStatus) == 'down'){
             $this->markTestSkipped('Prelive is not available');
@@ -1906,22 +1908,17 @@ class BatchRequestFunctionalTest extends \PHPUnit_Framework_TestCase
         $response = $resp->getXmlReader()->getAttribute("response");
         $this->assertEquals("Valid Format", $message);
         $this->assertEquals(0, $response);
-    }
+    }*/
 
-    public function tearDown()
-    {
-        $files = glob($this->direct . '/*'); // get all file names
-        foreach ($files as $file) { // iterate files
-            if (is_file($file))
-                unlink($file); // delete file
-        }
-        rmdir($this->direct);
-    }
-    public function test_addVendorDebit_with_vendorAddress()
+    public function test_addVendorDebit_with_vendorAddress_batchSFTP()
     {
         if(strtolower($this->preliveStatus) == 'down'){
             $this->markTestSkipped('Prelive is not available');
         }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
 
         $hash_in = array('id' => 'id',
             'fundingSubmerchantId' => '2111',
@@ -1944,21 +1941,27 @@ class BatchRequestFunctionalTest extends \PHPUnit_Framework_TestCase
                 'country' => 'US'),
 
         );
-        $batch_request = new BatchRequest ($this->direct);
-        $batch_request->addVendorDebit($hash_in);
 
-        $this->assertTrue(file_exists($batch_request->batch_file));
-        $this->assertEquals(1, $batch_request->total_txns);
+        $batch->addVendorDebit($hash_in);
 
-        $cts = $batch_request->getCountsAndAmounts();
-        $this->assertEquals(1, $cts ['vendorDebit'] ['count']);
-        $this->assertEquals(13, $cts ['vendorDebit'] ['amount']);
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
     }
-    public function test_addVendorCredit_with_vendorAddress()
+    public function test_addVendorCredit_with_vendorAddress_batchSFTP()
     {
         if(strtolower($this->preliveStatus) == 'down'){
             $this->markTestSkipped('Prelive is not available');
         }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
 
         $hash_in = array('id' => 'id',
             'fundingSubmerchantId' => '2111',
@@ -1982,15 +1985,561 @@ class BatchRequestFunctionalTest extends \PHPUnit_Framework_TestCase
                 'country' => 'US'),
 
         );
-        $batch_request = new BatchRequest ($this->direct);
-        $batch_request->addVendorCredit($hash_in);
 
-        $this->assertTrue(file_exists($batch_request->batch_file));
-        $this->assertEquals(1, $batch_request->total_txns);
+        $batch->addVendorCredit($hash_in);
 
-        $cts = $batch_request->getCountsAndAmounts();
-        $this->assertEquals(1, $cts ['vendorCredit'] ['count']);
-        $this->assertEquals(13, $cts ['vendorCredit'] ['amount']);
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+    }
+
+    public function test_auth_with_additionalCOFData_batch()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array(
+            'id'=>'0001',
+            'orderId' => '82364_cnpApiAuth',
+            'amount' => '2870',
+            'orderSource' => 'telephone',
+            'billToAddress' => array(
+                'name' => 'David Berman A',
+                'addressLine1' => '10 Main Street',
+                'city' => 'San Jose',
+                'state' => 'ca',
+                'zip' => '95032',
+                'country' => 'USA',
+                'email' => 'dberman@phoenixProcessing.com',
+                'phone' => '781-270-1111',
+                'sellerId' => '21234234A1',
+                'url' => 'www.google.com',
+            ),
+            'shipToAddress' => array(
+                'name' => 'Raymond J. Johnson Jr. B',
+                'addressLine1' => '123 Main Street',
+                'city' => 'McLean',
+                'state' => 'VA',
+                'zip' => '22102',
+                'country' => 'USA',
+                'email' => 'ray@rayjay.com',
+                'phone' => '978-275-0000',
+                'sellerId' => '21234234A2',
+                'url' => 'www.google.com',
+            ),
+            'crypto' => 'true',
+            'retailerAddress' => array(
+                'name' => 'John doe',
+                'addressLine1' => '123 Main Street',
+                'addressLine2' => '123 Main Street',
+                'addressLine3' => '123 Main Street',
+                'city' => 'Cincinnati',
+                'state' => 'OH',
+                'zip' => '45209',
+                'country' => 'USA',
+                'email' => 'noone@abc.com',
+                'phone' => '1234562783',
+                'sellerId' => '21234234A',
+                'companyName' => 'Google INC',
+                'url' => 'www.google.com',
+            ),
+            'additionalCOFData' => array(
+                'totalPaymentCount' => 'ND',
+                'paymentType' => 'Fixed Amount',
+                'uniqueId' => '234GTYH654RF13',
+                'frequencyOfMIT' => 'Annually',
+                'validationReference' => 'ANBH789UHY564RFC@EDB',
+                'sequenceIndicator' => '86',
+            ),
+            'card' => array(
+                'type' => 'VI',
+                'number' => '4005518220000002',
+                'expDate' => '0150',
+                'cardValidationNum' => '987',
+            ),
+            'orderChannel' => 'IN_STORE_KIOSK',
+            'fraudCheckStatus' => 'CLOSE',
+            'businessIndicator' => 'consumerBillPayment'
+
+        );
+
+        $batch->addAuth($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+
+    }
+
+    public function test_captureGivenAuth_with_additionalCOFData_batchSFTP()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+
+        $hash_in = array('id' => 'id',
+            'orderId' => '12344',
+            'amount' => '106',
+            'authInformation' => array(
+                'authDate' => '2002-10-09', 'authCode' => '543216',
+                'authAmount' => '12345'),
+            'billToAddress' => array('name' => 'Bob', 'city' => 'lowell', 'state' => 'MA', 'email' => 'vantiv.com'),
+            'processingInstructions' => array('bypassVelocityCheck' => 'true'),
+            'orderSource' => 'ecommerce',
+            'card' => array(
+                'type' => 'VI',
+                'number' => '4100000000000000',
+                'expDate' => '1210'),
+            'crypto' => 'true',
+            'billToAddress' => array(
+                'name' => 'David Berman A',
+                'addressLine1' => '10 Main Street',
+                'city' => 'San Jose',
+                'state' => 'ca',
+                'zip' => '95032',
+                'country' => 'USA',
+                'email' => 'dberman@phoenixProcessing.com',
+                'phone' => '781-270-1111',
+                'sellerId' => '21234234A1',
+                'url' => 'www.google.com',
+            ),
+            'shipToAddress' => array(
+                'name' => 'Raymond J. Johnson Jr. B',
+                'addressLine1' => '123 Main Street',
+                'city' => 'McLean',
+                'state' => 'VA',
+                'zip' => '22102',
+                'country' => 'USA',
+                'email' => 'ray@rayjay.com',
+                'phone' => '978-275-0000',
+                'sellerId' => '21234234A2',
+                'url' => 'www.google.com',
+            ),
+            'retailerAddress' => array(
+                'name' => 'John doe',
+                'addressLine1' => '123 Main Street',
+                'addressLine2' => '123 Main Street',
+                'addressLine3' => '123 Main Street',
+                'city' => 'Cincinnati',
+                'state' => 'OH',
+                'zip' => '45209',
+                'country' => 'USA',
+                'email' => 'noone@abc.com',
+                'phone' => '1234562783',
+                'sellerId' => '21234234A',
+                'companyName' => 'Google INC',
+                'url' => 'www.google.com',
+            ),
+            'additionalCOFData' => array(
+                'totalPaymentCount' => 'ND',
+                'paymentType' => 'Fixed Amount',
+                'uniqueId' => '234GTYH654RF13',
+                'frequencyOfMIT' => 'Annually',
+                'validationReference' => 'ANBH789UHY564RFC@EDB',
+                'sequenceIndicator' => '86',
+            ),
+            'businessIndicator' => 'buyOnlinePickUpInStore',
+        );
+        $batch->addCaptureGivenAuth($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+    }
+
+    public function test_sale_with_additionalCOFData_batchSFTP()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array(
+            'card' => array('type' => 'VI',
+                'number' => '4100000000000000',
+                'expDate' => '1213',
+                'cardValidationNum' => '1213'),
+            'id' => '1211',
+            'orderId' => '2111',
+            'reportGroup' => 'Planets',
+            'orderSource' => 'ecommerce',
+            'amount' => '123',
+            'merchantCategoryCode' => '6770',
+            'billToAddress' => array(
+                'name' => 'David Berman A',
+                'addressLine1' => '10 Main Street',
+                'city' => 'San Jose',
+                'state' => 'ca',
+                'zip' => '95032',
+                'country' => 'USA',
+                'email' => 'dberman@phoenixProcessing.com',
+                'phone' => '781-270-1111',
+                'sellerId' => '21234234A1',
+                'url' => 'www.google.com',
+            ),
+            'shipToAddress' => array(
+                'name' => 'Raymond J. Johnson Jr. B',
+                'addressLine1' => '123 Main Street',
+                'city' => 'McLean',
+                'state' => 'VA',
+                'zip' => '22102',
+                'country' => 'USA',
+                'email' => 'ray@rayjay.com',
+                'phone' => '978-275-0000',
+                'sellerId' => '21234234A2',
+                'url' => 'www.google.com',
+            ),
+            'crypto' => 'true',
+            'retailerAddress' => array(
+                'name' => 'John doe',
+                'addressLine1' => '123 Main Street',
+                'addressLine2' => '123 Main Street',
+                'addressLine3' => '123 Main Street',
+                'city' => 'Cincinnati',
+                'state' => 'OH',
+                'zip' => '45209',
+                'country' => 'USA',
+                'email' => 'noone@abc.com',
+                'phone' => '1234562783',
+                'sellerId' => '21234234A',
+                'companyName' => 'Google INC',
+                'url' => 'www.google.com',
+            ),
+            'additionalCOFData' => array(
+                'totalPaymentCount' => 'ND',
+                'paymentType' => 'Fixed Amount',
+                'uniqueId' => '234GTYH654RF13',
+                'frequencyOfMIT' => 'Annually',
+                'validationReference' => 'ANBH789UHY564RFC@EDB',
+                'sequenceIndicator' => '86',
+            ),
+            'orderChannel' => 'IN_STORE_KIOSK',
+            'fraudCheckStatus' => 'CLOSE',
+            'businessIndicator' => 'buyOnlinePickUpInStore',
+
+        );
+        $batch->addSale($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+    }
+
+    public function test_sale_customerInfo_with_accountUsername_batchSFTP()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array('merchantId' => '101', 'id' => '1211',
+            'version' => '12.24',
+            'reportGroup' => 'Planets',
+            'orderId' => '12344',
+            'amount' => '106',
+            'orderSource' => 'ecommerce',
+            'customerInfo' => array(
+                'incomeAmount' => '12345',
+                'incomeCurrency' => 'USD',
+                'yearsAtResidence' => '2',
+                'accountUsername' => 'Woolfoo',
+                'userAccountNumber' => '123456',
+                'userAccountEmail' => 'woolfoo@gmail.com',
+                'membershipId' => 'Member01',
+                'membershipPhone' => '9765431234',
+                'membershipEmail' => 'mem@abc.com',
+                'membershipName' => 'memName',
+                'accountCreatedDate' => '2022-04-04',
+                'userAccountPhone' => '123456789',
+            ),
+            'card' => array(
+                'type' => 'VI',
+                'number' => '4100000000000000',
+                'expDate' => '1210'
+            ));
+
+        $batch->addAuth($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+
+
+    }
+
+    public function test_enhancedData_with_discountCode_batchSFTP()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array('id' => 'id',
+            'card' => array('type' => 'VI',
+                'number' => '4100000000000000',
+                'expDate' => '1213',
+                'cardValidationNum' => '1213'),
+            'id' => '1211',
+            'orderId' => '22@33',
+            'reportGroup' => 'Planets',
+            'orderSource' => 'ecommerce',
+            'amount' => '0',
+            'enhancedData' => array(
+                'salesTax' => '500',
+                'taxExempt' => false,
+                'detailTax0' => array(
+                    'taxAmount' => '200',
+                    'taxRate' => '0.06',
+                    'taxIncludedInTotal' => true
+                ),
+                'detailTax1' => array(
+                    'taxAmount' => '300',
+                    'taxRate' => '0.10',
+                    'taxIncludedInTotal' => true
+                ),'lineItemData0' => array(
+                    'itemSequenceNumber' => '1',
+                    'itemDescription' => 'product 1',
+                    'productCode' => '123',
+                    'quantity' => 3,
+                    'unitOfMeasure' => 'unit',
+                    'taxAmount' => 200,
+                    'detailTax' => array(
+                        'taxIncludedInTotal' => true,
+                        'taxAmount' => 200
+                    ),
+                    'itemCategory' => 'Aparel',
+                    'itemSubCategory' => 'Clothing',
+                    'productId' => '1001',
+                    'productName' => 'N1',
+                ),
+                'lineItemData1' => array(
+                    'itemSequenceNumber' => '2',
+                    'itemDescription' => 'product 2',
+                    'productCode' => '456',
+                    'quantity' => 1,
+                    'unitOfMeasure' => 'unit',
+                    'taxAmount' => 300,
+                    'detailTax' => array(
+                        'taxIncludedInTotal' => true,
+                        'taxAmount' => 300
+                    )
+                ),
+                'discountCode' => 'oneTimeDis',
+                'discountPercent' => '12',
+                'fulfilmentMethodType' => 'COUNTER_PICKUP'
+            )
+
+        );
+        $batch->addAuth($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+
+    }
+
+    public function test_simple_sale_with_interac_batchSFTP()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array(
+            'card' => array('type' => 'IC',
+                'number' => '4100000000000000',
+                'expDate' => '1213',
+                'cardValidationNum' => '1213'),
+            'cardholderAuthentication' => array(
+                /// base64 value for dummy number '123456789012345678901234567890123456789012345678901234567890'
+                /// System should accept the request with length 60 of authenticationValueType
+                'authenticationValue' => 'MTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkwMTIzNDU2Nzg5MDEyMzQ1Njc4OTAxMjM0NTY3ODkw'
+            ),
+            'id' => '1211',
+            'orderId' => '2111',
+            'reportGroup' => 'Planets',
+            'orderSource' => 'ecommerce',
+            'amount' => '123',
+            'billToAddress' => array(
+                'name' => 'David Berman A',
+                'addressLine1' => '10 Main Street',
+                'city' => 'San Jose',
+                'state' => 'ca',
+                'zip' => '95032',
+                'country' => 'USA',
+                'email' => 'dberman@phoenixProcessing.com',
+                'phone' => '781-270-1111',
+                'sellerId' => '21234234A1',
+                'url' => 'www.google.com',
+            ))
+        ;
+
+        $batch->addSale($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+    }
+
+    public function test_FastAcessFund_with_cardholderAddress_batchSFPT()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+
+        $hash_in = array('id' => 'id',
+            'fundingSubmerchantId' => '2111',
+            'submerchantName' => '001',
+            'fundsTransferId' => '1234567891111111',
+            'amount' => '13',
+            'card' => array(
+                'type' => 'VI',
+                'number' => '4100000000000000',
+                'expDate' => '1210'
+            ),
+            'cardholderAddress' => array(
+                'addressLine1' => '2 Main St.',
+                'addressLine2' => 'Apt. 222',
+                'addressLine3' => 'NA',
+                'city' => 'Riverside',
+                'state' => 'RI',
+                'zip' => '02915',
+                'country' => 'US')
+        );
+        $batch->addFastAccessFunding($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+
+    }
+
+    public function test_simple_credit_with_pin_and_optional_order_id()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array(
+            'cnpTxnId' => '12312312',
+            'orderId' => '22@33123456789012345678901234567890',
+            'id' => 'id',
+            'reportGroup' => 'Planets',
+            'amount' => '123',
+            'secondaryAmount' => '3214',
+            'surchargeAmount' => '1',
+            'pin' => '3333'
+        );
+
+        $batch->addCredit($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+    }
+
+    public function test_simple_capture_with_optional_order_id()
+    {
+        if(strtolower($this->preliveStatus) == 'down'){
+            $this->markTestSkipped('Prelive is not available');
+        }
+
+        $request = new CnpRequest();
+
+        $batch = new BatchRequest();
+        $hash_in = array('id' => 'id',
+            'cnpTxnId' => '1234567891234567891',
+            'orderId' => '22@33123456789012345678901234567890',
+            'amount' => '123');
+
+        $batch->addCapture($hash_in);
+
+        $request->addBatchRequest($batch);
+
+        $resp = new CnpResponseProcessor($request->sendToCnp());
+
+        $message = $resp->getXmlReader()->getAttribute("message");
+        $response = $resp->getXmlReader()->getAttribute("response");
+        $this->assertEquals("Valid Format", $message);
+        $this->assertEquals(0, $response);
+    }
+
+    public function tearDown()
+    {
+        $files = glob($this->direct . '/*'); // get all file names
+        foreach ($files as $file) { // iterate files
+            if (is_file($file))
+                unlink($file); // delete file
+        }
+        rmdir($this->direct);
     }
 
 }