+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 Note: It contains changes from cnpAPI v12.19, v12.20, v12.21, v12.22 & 12.23. In case you need any feature supported by cnpAPI v12.19, v12.20, v12.21, v12.22 or v12.23, please use SDK version 12.24.0.
 
 * Feature: [cnpAPI v12.24] New Enum value highRiskSecuritiesPurchase, fundTransfer and walletTransfer included in businessIndicatorEnum.
-* Feature: [cnpAPI v12.24] New element fraudCheckStatus is added for authorization, captureGivenAuth and  sale transactions.
+* Feature: [cnpAPI v12.24] New element fraudCheckStatus is added for authorization and sale transactions.
 * Feature: [cnpAPI v12.24] New element crypto is added for authorization, captureGivenAuth and  sale transactions.
 * Feature: [cnpAPI v12.23] New Enum value buyOnlinePickUpInStore included in businessIndicatorEnum.
 * Feature: [cnpAPI v12.23] sellerId & url elements are added for new element retailerAddress and for existing contact type elements.